Split the `seq_join` file into separate modules
One for local spawn and one for multi-threading
